oncall/requirements.txt
# On-Call Troubleshooting Agent - Python Dependencies
# ArtemisHealth DevOps Engineering

# Core Claude Agent SDK
claude-agent-sdk>=0.1.0

# Anthropic API (for daemon mode LLM calls)
anthropic>=0.39.0

# Kubernetes Integration
kubernetes>=29.0.0
pyyaml>=6.0.1

# GitHub Integration
PyGithub>=2.1.1

# AWS Integration
boto3>=1.34.0

# Datadog Integration
datadog-api-client>=2.28.0

# Async HTTP Requests
aiohttp>=3.9.0

# FastAPI and API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
slowapi>=0.1.9

# Environment Management
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=24.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0

# Development Tools
ipython>=8.18.0
ipdb>=0.13.13